fix use url test & implmentation
