+++ .changeset/chatty-cameras-wash.md
@@ -1,0 +1,5 @@
+---
+'@shopify/hydrogen': patch
+---
+
+Fix useUrl() when it is in RSC mode
